Move Regobs enums to separate file
This decouples the alternatives for registrations from the logic
itself. There have been suggestions to move the enums to a separate
package. Either way, it's good to have them separate from the logic.
